=Dao tests repaired, package structure repaired, mvn clean install repaired
