+++ referenceManager-api/pom.xml
@@ -15,19 +15,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>cz.muni.fi.pa165.referenceManager</groupId>
-            <artifactId>referenceManager-service</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-        <dependency>
             <groupId>javax.inject</groupId>
             <artifactId>javax.inject</artifactId>
             <version>1</version>
-        </dependency>
-        <dependency>
-            <groupId>cz.muni.fi.pa165.referenceManager</groupId>
-            <artifactId>referenceManager-persistence</artifactId>
-            <version>${project.parent.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework</groupId>
