--- referenceManager-api/src/main/java/cz/fi/muni/pa165/referenceManager/dto/NoteDTO.java
@@ -1,4 +1,0 @@
-package cz.fi.muni.pa165.referenceManager.dto;
-
-public interface NoteDTO {
-}
