--- referenceManager-api/src/main/java/cz/fi/muni/pa165/referenceManager/dto/TagDTO.java
@@ -1,4 +1,0 @@
-package cz.fi.muni.pa165.referenceManager.dto;
-
-public interface TagDTO {
-}
