--- referenceManager-api/src/main/java/cz/fi/muni/pa165/referenceManager/dto/UserLoginDTO.java
@@ -1,24 +1,0 @@
-package cz.fi.muni.pa165.referenceManager.dto;
-
-public class UserLoginDTO {
-
-    private String email;
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-}
