+++ referenceManager-api/src/main/java/cz/muni/fi/pa165/referenceManager/dto/NoteDTO.java
@@ -1,0 +1,4 @@
+package cz.muni.fi.pa165.referenceManager.dto;
+
+public class NoteDTO {
+}
