+++ referenceManager-api/src/main/java/cz/muni/fi/pa165/referenceManager/facade/ReferenceFacade.java
@@ -1,0 +1,16 @@
+package cz.muni.fi.pa165.referenceManager.facade;
+
+import cz.muni.fi.pa165.referenceManager.dto.NoteDTO;
+import cz.muni.fi.pa165.referenceManager.dto.ReferenceDTO;
+
+import java.util.List;
+
+public interface ReferenceFacade {
+    void createReference(ReferenceDTO r);
+    void updateReference(ReferenceDTO r);
+    void deleteReference(Long id);
+    List<ReferenceDTO> getAllReferences();
+    ReferenceDTO getReferenceById(Long id);
+    void addNote(Long referenceId, NoteDTO note);
+    void removeNote(Long referenceId, NoteDTO note);
+}
