--- referenceManager-api/src/test/java/cz/fi/muni/pa165/referenceManager/ReferenceFacadeTest.java
@@ -1,4 +1,0 @@
-package cz.fi.muni.pa165.referenceManager;
-
-public class ReferenceFacadeTest {
-}
