+++ referenceManager-api/src/test/java/cz/muni/fi/pa165/referenceManager/ReferenceFacadeTest.java
@@ -1,0 +1,4 @@
+package cz.muni.fi.pa165.referenceManager;
+
+public class ReferenceFacadeTest {
+}
