+++ referenceManager-persistence/src/test/java/cz/muni/fi/pa165/referenceManager/dao/ReferenceDaoTest.java
@@ -29,7 +29,7 @@ public class ReferenceDaoTest {
     @PersistenceContext
     private EntityManager em;
 
-    @Test(expected = PersistenceException.class)
+    @Test(expected = NullPointerException.class)
     public void testNewReferenceWithoutTitleError() {
         Reference reference = new Reference();
         referenceDao.create(reference);
