+++ referenceManager-service/pom.xml
@@ -21,11 +21,6 @@
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
-            <groupId>cz.muni.fi.pa165.referenceManager</groupId>
-            <artifactId>referenceManager-api</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-        <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
             <version>5.0.1.RELEASE</version>
