--- referenceManager-service/src/main/java/cz/fi/muni/pa165/referenceManager/service/MappingService.java
@@ -1,15 +1,0 @@
-package cz.fi.muni.pa165.referenceManager.service;
-
-import org.dozer.Mapper;
-
-import java.util.Collection;
-import java.util.List;
-
-public interface MappingService {
-
-    public <T> List<T> mapTo(Collection<?> objects, Class<T> mapToClass);
-
-    public <T> T mapTo(Object u, Class<T> mapToClass);
-
-    public Mapper getMapper();
-}
